Added grenade explosion WIP, bullets are now synched better, grenade synch needs work.

+++ src/game_time.cpp
@@ -2,7 +2,7 @@
 
 namespace arena
 {
-	GameTime::GameTime(const float32 delta, const float64 total) : m_delta(delta),
+	GameTime::GameTime(const float64 delta, const float64 total) : m_delta(delta),
 																   m_total(total)
 	{
 	}
