+++ src/game_time.h
@@ -7,10 +7,10 @@ namespace arena
 	class GameTime final
 	{
 	public:
-		const float32 m_delta;
+		const float64 m_delta;
 		const float64 m_total;
 
-		GameTime(const float32 delta, const float64 total);
+		GameTime(const float64 delta, const float64 total);
 
 		~GameTime() = default;
 	};
